+++ src/pieces/commands/__init__.py
@@ -1,11 +1,11 @@
-from .cli_loop import loop
+from .cli_loop import CliLoop
 from .change_model import change_model
 from .search_command import search
 from .list_command import ListCommand
 from .version_command import version
 from .signout_command import sign_out
 
-__all__ = ['loop',
+__all__ = ['CliLoop',
            'version',
            'search',
            'change_model',